Separate multiple experiments (#10)
* Add function to separate main sections with multiple experiments

Added this function to stopadforms instead of synapseforms since this is really only relevant here. Add documentation.

* Use tidier data

Use tidier data from the start, which helps with separating out the main sections with multiple functions. Use synapseforms to get a tidier version of the data and then the new cleaning function to separate out the multiple experiments. Update synapseforms calls to include is_tidier parameter, when necessary.

* Update packages (#11)

* Remove outdated is_tidier param from synapseforms calls