Use inner join in pull_reviews_table
It's possible to have mismatches between submission data and reviews (i.e. the
function can take a data frame of just a single submission, but the table query
would return all the reviews for all submissions). This causes errors when we
go to calculate the weighted scores by section, since there could be missing
data in the rows where we have reviews but haven't passed in the submission data
for that submission.